Unit Test and merg request some changes
