+++ AccountsApi.Tests/GlobalSuppressions.cs
@@ -5,3 +5,4 @@
 
 using System.Diagnostics.CodeAnalysis;
 
+[assembly: SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "<Pending>", Scope = "member", Target = "~M:AccountsApi.Tests.V1.Controllers.AccountsApiControllerTests.GetAllArrears_UseCaseThrowException_Returns500~System.Threading.Tasks.Task")]
