+++ AccountsApi.Tests/V1/Controllers/AccountsApiControllerTests.cs
@@ -11,6 +11,7 @@ using AccountsApi.V1.Boundary.Request;
 using AccountsApi.V1.Boundary.Response;
 using AccountsApi.V1.Controllers;
 using AccountsApi.V1.Domain;
+using AccountsApi.V1.Infrastructure;
 using AccountsApi.V1.UseCase.Interfaces;
 using AutoFixture;
 using FluentAssertions;
@@ -27,13 +28,12 @@ namespace AccountsApi.Tests.V1.Controllers
     public class AccountsApiControllerTests
     {
         private readonly AccountApiController _sut;
-        private readonly ControllerContext _controllerContext;
-        private readonly HttpContext _httpContext;
 
         private readonly Mock<IGetAllUseCase> _getAllUseCase;
         private readonly Mock<IGetByIdUseCase> _getByIdUseCase;
         private readonly Mock<IAddUseCase> _addUseCase;
         private readonly Mock<IUpdateUseCase> _updateUseCase;
+        private readonly Mock<IGetAllArrearsUseCase> _getAllArrearsUseCase;
 
         private readonly Fixture _fixture = new Fixture();
 
@@ -44,13 +44,14 @@ namespace AccountsApi.Tests.V1.Controllers
             _getByIdUseCase = new Mock<IGetByIdUseCase>();
             _addUseCase = new Mock<IAddUseCase>();
             _updateUseCase = new Mock<IUpdateUseCase>();
+            _getAllArrearsUseCase = new Mock<IGetAllArrearsUseCase>();
 
-            _httpContext = new DefaultHttpContext();
-            _controllerContext = new ControllerContext(new ActionContext(_httpContext, new RouteData(), new ControllerActionDescriptor()));
+            HttpContext httpContext = new DefaultHttpContext();
+            var controllerContext = new ControllerContext(new ActionContext(httpContext, new RouteData(), new ControllerActionDescriptor()));
             _sut = new AccountApiController(_getAllUseCase.Object, _getByIdUseCase.Object,
-                _addUseCase.Object, _updateUseCase.Object)
+                _addUseCase.Object, _updateUseCase.Object, _getAllArrearsUseCase.Object)
             {
-                ControllerContext = _controllerContext
+                ControllerContext = controllerContext
             };
         }
         [Fact]
@@ -271,9 +272,10 @@ namespace AccountsApi.Tests.V1.Controllers
 
             // Assert
             result.Should().BeOfType<BadRequestObjectResult>();
-        } 
+        }
         #endregion
 
+        #region Post
         [Fact]
         public async Task PostSuccessfullReturnsAccountModel()
         {
@@ -344,7 +346,134 @@ namespace AccountsApi.Tests.V1.Controllers
             model.AccountStatus.Should().Be(AccountStatus.Active);
             model.AccountBalance.Should().Be(0);
 
+        } 
+        #endregion
+
+        #region Arrears
+
+        [Fact]
+        public async Task GetAllArrears_UseCaseReturnsList_Returns200()
+        {
+            var useCaseResponse = new AccountResponses()
+            {
+                AccountResponseList = new List<AccountModel>()
+                {
+                    new AccountModel()
+                    {
+                        Id = new Guid("b3b91924-1a3d-44b7-b38a-ae4ae5e57b69"),
+                        TargetId = new Guid("2da59b6b-cdcb-46bd-ac61-1c10d1046285"),
+                        StartDate = new DateTime(2021, 7, 1),
+                        CreatedDate = new DateTime(2021, 6, 1),
+                        EndDate = new DateTime(2021, 9, 1),
+                        LastUpdatedDate = new DateTime(2021, 7, 1),
+                        AccountBalance = 100.0M,
+                        AccountStatus = AccountStatus.Active,
+                        AccountType = AccountType.Master,
+                        AgreementType = "string",
+                        CreatedBy = "Admin",
+                        RentGroupType = RentGroupType.Garages,
+                        LastUpdatedBy = "Admin",
+                        TargetType = TargetType.Block
+                    },
+                    new AccountModel()
+                    {
+                        Id = new Guid("17b107e7-b7a1-4c14-a1c9-0630cebdf28e"),
+                        TargetId = new Guid("2da59b6b-cdcb-46bd-ac61-1c10d1046285"),
+                        StartDate = new DateTime(2021, 7, 1),
+                        CreatedDate = new DateTime(2021, 6, 1),
+                        EndDate = new DateTime(2021, 9, 1),
+                        LastUpdatedDate = new DateTime(2021, 7, 1),
+                        AccountBalance = 110.0M,
+                        AccountStatus = AccountStatus.Suspended,
+                        AccountType = AccountType.Master,
+                        AgreementType = "string",
+                        CreatedBy = "Admin",
+                        RentGroupType = RentGroupType.GenFundRents,
+                        LastUpdatedBy = "Admin",
+                        TargetType = TargetType.Core
+                    }
+                }
+            };
+
+            var request = new ArrearRequest()
+            {
+                Type = AccountType.Master,
+                SortBy = "AgreementType",
+                Direction = Direction.Asc
+            };
+
+            _getAllArrearsUseCase.Setup(_ => _.ExecuteAsync(It.IsAny<ArrearRequest>()))
+                .ReturnsAsync(useCaseResponse);
+
+            var result = await _sut.GetArrears(request).ConfigureAwait(false);
+
+            result.Should().NotBeNull();
+
+            var okResult = result as OkObjectResult;
+
+            okResult.Should().NotBeNull();
+
+            okResult.Value.Should().NotBeNull();
+
+            var response = okResult.Value as AccountResponses;
+
+            response.Should().NotBeNull();
+
+            response.AccountResponseList.Should().NotBeNull();
+
+            response.AccountResponseList.Should().HaveCount(2);
+
+            response.AccountResponseList[0].Should().BeEquivalentTo(useCaseResponse.AccountResponseList[0]);
+            response.AccountResponseList[1].Should().BeEquivalentTo(useCaseResponse.AccountResponseList[1]);
+
         }
 
+        [Fact]
+        public async Task GetAllArrears_NullRequestModel_Returns400()
+        {
+            ArrearRequest request = null;
+
+            var result = await _sut.GetArrears(request).ConfigureAwait(false);
+
+            result.Should().NotBeNull();
+
+            var badRequest = result as BadRequestObjectResult;
+
+            badRequest.Should().NotBeNull();
+
+            badRequest.Value.Should().NotBeNull();
+
+            var baseErrorResponse = badRequest.Value as BaseErrorResponse;
+
+            baseErrorResponse.Should().NotBeNull();
+
+            baseErrorResponse.StatusCode.Should().Be((int) HttpStatusCode.BadRequest);
+
+            baseErrorResponse.Message.Should().BeEquivalentTo("ArrearRequest can't be null");
+
+            baseErrorResponse.Details.Should().BeEquivalentTo(string.Empty);
+        }
+
+        [Fact]
+        public async Task GetAllArrears_UseCaseThrowException_Returns500()
+        {
+            var request = new ArrearRequest();
+
+            _getAllArrearsUseCase.Setup(_ => _.ExecuteAsync(It.IsAny<ArrearRequest>()))
+                .ThrowsAsync(new Exception("Test exception"));
+
+            try
+            {
+                var result = await _sut.GetArrears(new ArrearRequest()).ConfigureAwait(false);
+                AssertExtensions.Fail();
+            }
+            catch (Exception ex)
+            {
+                ex.GetType().Should().Be(typeof(Exception));
+                ex.Message.Should().Be("Test exception");
+            }
+        }
+
+        #endregion
     }
 }
