+++ AccountsApi/V1/Boundary/Request/ArrearRequest.cs
@@ -1,6 +1,6 @@
-using AccountApi.V1.Infrastructure;
 using AccountsApi.V1.Domain;
 using System.ComponentModel.DataAnnotations;
+using AccountsApi.V1.Infrastructure;
 
 namespace AccountsApi.V1.Boundary.Request
 {
